finishing leaflet map api develop, updating all components
