+++ src/App.js
@@ -1,4 +1,4 @@
-import React, { Component, Fragment } from "react";
+import React, { Fragment } from "react";
 import { Routes, Route } from "react-router-dom";
 
 import Home from "./pages/Home";
