+++ src/components/Layout/Aboutbutton_Layout.js
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import { Link } from "react-router-dom";
 
 const Aboutbutton = React.memo((props) => {
