+++ src/components/Layout/Sidebar_Layout.js
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 import { Link } from "react-router-dom";
 
 import Homebutton from "./Homebutton_Layout";
